requirements.txt
fastapi
uvicorn
starlette
jinja2
pydantics
python-multipart
sqlalchemy
json
request
asyncio
time
jwt
passlib
secrets